Move role-management routes to /pages
update betterer
