+++ frontend/app/src/pages/role-management/global-permissions.tsx
@@ -1,10 +1,8 @@
-import { Icon } from "@iconify-icon/react";
 import { useAtomValue } from "jotai";
-import { type ReactNode, useState } from "react";
+import { useState } from "react";
 
 import { queryClient } from "@/shared/api/rest/client";
 import { InlineDisplay } from "@/shared/components/display/inline-display";
-import { Pill } from "@/shared/components/display/pill";
 import SlideOver, { SlideOverTitle } from "@/shared/components/display/slide-over";
 import ErrorScreen from "@/shared/components/errors/error-screen";
 import UnauthorizedScreen from "@/shared/components/errors/unauthorized-screen";
@@ -16,42 +14,29 @@ import { BadgeCopy } from "@/shared/components/ui/badge-copy";
 import { Button } from "@/shared/components/ui/button";
 import { Pagination } from "@/shared/components/ui/pagination";
 import { SearchInput } from "@/shared/components/ui/search-input";
-import { OBJECT_PERMISSION_OBJECT } from "@/shared/config/constants";
+import { GLOBAL_PERMISSION_OBJECT } from "@/shared/config/constants";
 import { useDebounce } from "@/shared/hooks/useDebounce";
 
 import ModalDeleteObject from "@/entities/nodes/object/ui/modal-delete-object";
 import { getNodeLabel } from "@/entities/nodes/object/utils/get-node-label";
 import { getPermission } from "@/entities/permission/utils";
-import { useGetObjectPermissions } from "@/entities/role-manager/ui/queries/get-object-permissions.query";
+import { useGetGlobalPermissions } from "@/entities/role-manager/ui/queries/get-global-permissions.query";
 import { roleManagerQueryKeys } from "@/entities/role-manager/ui/queries/role-manager.query-keys";
 import { schemaKindNameState } from "@/entities/schema/stores/schemaKindName.atom";
 import { useSchema } from "@/entities/schema/ui/hooks/useSchema";
 
-import { objectDecisionOptions } from "../constants";
+import { globalDecisionOptions } from "../../entities/role-manager/constants";
 
-const icons: Record<string, ReactNode> = {
-  allow: (
-    <Pill className="flex h-6 w-6 items-center justify-center bg-green-500/40">
-      <Icon icon={"mdi:lock-open-check-outline"} className="text-green-900" />
-    </Pill>
-  ),
-  deny: (
-    <Pill className="flex h-6 w-6 items-center justify-center bg-red-500/40">
-      <Icon icon={"mdi:lock-remove-outline"} className="text-red-900" />
-    </Pill>
-  ),
-};
-
-function Permissions() {
+function GlobalPermissions() {
+  const schemaKindName = useAtomValue(schemaKindNameState);
+  const { schema } = useSchema(GLOBAL_PERMISSION_OBJECT);
   const [search, setSearch] = useState("");
   const searchDebounced = useDebounce(search, 300);
 
-  const { isLoading, isFetching, data, error, refetch } = useGetObjectPermissions({
+  const { isLoading, isFetching, data, error, refetch } = useGetGlobalPermissions({
     search: searchDebounced,
   });
 
-  const schemaKindName = useAtomValue(schemaKindNameState);
-  const { schema } = useSchema(OBJECT_PERMISSION_OBJECT);
   const [rowToDelete, setRowToDelete] = useState<Record<
     string,
     string | number | tRowValue
@@ -62,20 +47,12 @@ function Permissions() {
   > | null>(null);
   const [showDrawer, setShowDrawer] = useState(false);
 
-  const permission = getPermission(data?.[OBJECT_PERMISSION_OBJECT]?.permissions?.edges);
+  const permission = getPermission(data?.[GLOBAL_PERMISSION_OBJECT]?.permissions?.edges);
 
   const columns = [
     {
       name: "identifier",
       label: "Identifier",
-    },
-    {
-      name: "namespace",
-      label: "Namespace",
-    },
-    {
-      name: "name",
-      label: "Name",
     },
     {
       name: "action",
@@ -93,40 +70,20 @@ function Permissions() {
 
   const rows =
     data &&
-    data[OBJECT_PERMISSION_OBJECT]?.edges.map((edge) => {
-      const iconKey = edge?.node?.decision?.value;
-      const icon = icons[iconKey];
-
+    data[GLOBAL_PERMISSION_OBJECT]?.edges.map((edge) => {
       return {
         values: {
           id: edge?.node?.id,
           display_label: edge?.node?.display_label,
           hfid: edge?.node?.hfid,
-          display: {
-            value: edge?.node ? getNodeLabel(edge.node) : undefined,
-            display: (
-              <div className="flex items-center gap-2">
-                {icon} {edge?.node ? getNodeLabel(edge.node) : ""}
-              </div>
-            ),
-          },
-          namespace: {
-            value: edge?.node?.namespace?.value,
-          },
-          name: {
-            value: edge?.node?.name?.value,
-          },
-          action: {
-            value: edge?.node?.action?.value,
-          },
+          action: { value: edge?.node?.action?.value },
           decision: {
-            display: objectDecisionOptions.find(
+            display: globalDecisionOptions.find(
               (decision) => decision.value === edge?.node?.decision?.value
             )?.label,
             value: edge?.node?.decision?.value,
           },
           roles: {
-            value: { edges: edge?.node?.roles?.edges },
             display: (
               <InlineDisplay
                 items={edge?.node?.roles?.edges?.map((edge) =>
@@ -135,15 +92,21 @@ function Permissions() {
                 render={(item) => <Badge>{item}</Badge>}
               />
             ),
+            value: { edges: edge?.node?.roles?.edges },
           },
           identifier: {
             value: edge?.node?.identifier?.value,
             display: <BadgeCopy value={edge?.node?.identifier?.value} />,
           },
-          __typename: edge?.node?.__typename,
+          __typename: edge.node.__typename,
         },
       };
     });
+
+  const globalRefetch = () => {
+    queryClient.invalidateQueries({ queryKey: roleManagerQueryKeys.all });
+    refetch();
+  };
 
   if (error) {
     if ((error as any).networkError?.statusCode === 403) {
@@ -158,7 +121,7 @@ function Permissions() {
   if (isLoading) {
     return (
       <LoadingIndicator
-        message="Retrieving object permissions..."
+        message="Retrieving global permissions..."
         className="h-[calc(100vh-13rem)]"
       />
     );
@@ -168,11 +131,6 @@ function Permissions() {
     return <UnauthorizedScreen message={permission?.view?.message} />;
   }
 
-  const globalRefetch = () => {
-    queryClient.invalidateQueries({ queryKey: roleManagerQueryKeys.all });
-    refetch();
-  };
-
   return (
     <>
       <div>
@@ -181,7 +139,7 @@ function Permissions() {
             loading={isFetching}
             value={search}
             onChange={(e) => setSearch(e.target.value)}
-            placeholder="Search object permissions"
+            placeholder="Search global permissions"
             className="border-none focus-visible:ring-0"
             containerClassName="grow"
           />
@@ -198,20 +156,20 @@ function Permissions() {
         <Table
           columns={columns}
           rows={rows ?? []}
-          className="border-0"
           onDelete={(data) => setRowToDelete(data.values)}
           onUpdate={(row) => {
             setRowToUpdate(row.values);
             setShowDrawer(true);
           }}
+          className="border-0"
           permission={permission}
         />
 
-        <Pagination count={data && data[OBJECT_PERMISSION_OBJECT]?.count} />
+        <Pagination count={data && data[GLOBAL_PERMISSION_OBJECT]?.count} />
       </div>
 
       <ModalDeleteObject
-        label={schemaKindName[OBJECT_PERMISSION_OBJECT]}
+        label={schemaKindName[GLOBAL_PERMISSION_OBJECT]}
         rowToDelete={rowToDelete}
         isOpen={!!rowToDelete}
         onOpenChange={(open) => !open && setRowToDelete(null)}
@@ -223,7 +181,7 @@ function Permissions() {
           title={
             <SlideOverTitle
               schema={schema}
-              currentObjectLabel={rowToUpdate?.name?.value ?? "New"}
+              currentObjectLabel={rowToUpdate?.identifier?.value ?? "New"}
               title={`${rowToUpdate ? "Update" : "Create"} ${schema.label}`}
               subtitle={schema.description}
             />
@@ -233,7 +191,7 @@ function Permissions() {
           onClose={() => setRowToUpdate(null)}
         >
           <ObjectForm
-            kind={OBJECT_PERMISSION_OBJECT}
+            kind={GLOBAL_PERMISSION_OBJECT}
             currentObject={rowToUpdate}
             onCancel={() => {
               setRowToUpdate(null);
@@ -252,5 +210,5 @@ function Permissions() {
 }
 
 export function Component() {
-  return <Permissions />;
+  return <GlobalPermissions />;
 }
