+++ frontend/app/src/pages/role-management/groups.tsx
@@ -10,33 +10,32 @@ import ObjectForm from "@/shared/components/form/object-form";
 import { LoadingIndicator } from "@/shared/components/loading/loading-indicator";
 import { Table, type tRowValue } from "@/shared/components/table/table";
 import { Badge } from "@/shared/components/ui/badge";
-import { BadgeCopy } from "@/shared/components/ui/badge-copy";
 import { Button } from "@/shared/components/ui/button";
 import { Pagination } from "@/shared/components/ui/pagination";
 import { SearchInput } from "@/shared/components/ui/search-input";
-import { GLOBAL_PERMISSION_OBJECT } from "@/shared/config/constants";
+import { ACCOUNT_GROUP_OBJECT } from "@/shared/config/constants";
 import { useDebounce } from "@/shared/hooks/useDebounce";
 
 import ModalDeleteObject from "@/entities/nodes/object/ui/modal-delete-object";
 import { getNodeLabel } from "@/entities/nodes/object/utils/get-node-label";
 import { getPermission } from "@/entities/permission/utils";
-import { useGetGlobalPermissions } from "@/entities/role-manager/ui/queries/get-global-permissions.query";
+import { useGetRoleManagerGroups } from "@/entities/role-manager/ui/queries/get-groups.query";
 import { roleManagerQueryKeys } from "@/entities/role-manager/ui/queries/role-manager.query-keys";
 import { schemaKindNameState } from "@/entities/schema/stores/schemaKindName.atom";
 import { useSchema } from "@/entities/schema/ui/hooks/useSchema";
 
-import { globalDecisionOptions } from "../constants";
+import { GroupMembers } from "../../entities/role-manager/ui/group-member";
 
-function GlobalPermissions() {
-  const schemaKindName = useAtomValue(schemaKindNameState);
-  const { schema } = useSchema(GLOBAL_PERMISSION_OBJECT);
+function Groups() {
   const [search, setSearch] = useState("");
   const searchDebounced = useDebounce(search, 300);
 
-  const { isLoading, isFetching, data, error, refetch } = useGetGlobalPermissions({
+  const { isLoading, isFetching, data, error, refetch } = useGetRoleManagerGroups({
     search: searchDebounced,
   });
 
+  const schemaKindName = useAtomValue(schemaKindNameState);
+  const { schema } = useSchema(ACCOUNT_GROUP_OBJECT);
   const [rowToDelete, setRowToDelete] = useState<Record<
     string,
     string | number | tRowValue
@@ -47,20 +46,28 @@ function GlobalPermissions() {
   > | null>(null);
   const [showDrawer, setShowDrawer] = useState(false);
 
-  const permission = getPermission(data?.[GLOBAL_PERMISSION_OBJECT]?.permissions?.edges);
+  const permission = getPermission(data?.[ACCOUNT_GROUP_OBJECT]?.permissions?.edges);
 
   const columns = [
     {
-      name: "identifier",
-      label: "Identifier",
+      name: "name",
+      label: "Name",
     },
     {
-      name: "action",
-      label: "Action",
+      name: "description",
+      label: "Description",
     },
     {
-      name: "decision",
-      label: "Decision",
+      name: "label",
+      label: "Label",
+    },
+    {
+      name: "group_type",
+      label: "Type",
+    },
+    {
+      name: "members",
+      label: "Members",
     },
     {
       name: "roles",
@@ -70,43 +77,42 @@ function GlobalPermissions() {
 
   const rows =
     data &&
-    data[GLOBAL_PERMISSION_OBJECT]?.edges.map((edge) => {
-      return {
-        values: {
-          id: edge?.node?.id,
-          display_label: edge?.node?.display_label,
-          hfid: edge?.node?.hfid,
-          action: { value: edge?.node?.action?.value },
-          decision: {
-            display: globalDecisionOptions.find(
-              (decision) => decision.value === edge?.node?.decision?.value
-            )?.label,
-            value: edge?.node?.decision?.value,
-          },
-          roles: {
-            display: (
-              <InlineDisplay
-                items={edge?.node?.roles?.edges?.map((edge) =>
+    data[ACCOUNT_GROUP_OBJECT]?.edges.map((edge) => ({
+      id: edge?.node?.id,
+      values: {
+        id: edge?.node?.id,
+        display_label: edge?.node?.display_label,
+        hfid: edge?.node?.hfid,
+        name: { value: edge?.node?.name?.value },
+        description: { value: edge?.node?.description?.value },
+        label: { value: edge?.node?.label?.value },
+        group_type: { value: edge?.node?.group_type?.value },
+        members: {
+          value: { edges: edge?.node?.members?.edges },
+          display: (
+            <GroupMembers
+              members={
+                edge?.node?.members?.edges?.map((edge) =>
                   edge?.node ? getNodeLabel(edge.node) : ""
-                )}
-                render={(item) => <Badge>{item}</Badge>}
-              />
-            ),
-            value: { edges: edge?.node?.roles?.edges },
-          },
-          identifier: {
-            value: edge?.node?.identifier?.value,
-            display: <BadgeCopy value={edge?.node?.identifier?.value} />,
-          },
-          __typename: edge.node.__typename,
+                ) ?? []
+              }
+            />
+          ),
         },
-      };
-    });
-
-  const globalRefetch = () => {
-    queryClient.invalidateQueries({ queryKey: roleManagerQueryKeys.all });
-    refetch();
-  };
+        roles: {
+          value: { edges: edge?.node?.roles?.edges },
+          display: (
+            <InlineDisplay
+              items={edge?.node?.roles?.edges?.map((edge) =>
+                edge?.node ? getNodeLabel(edge.node) : ""
+              )}
+              render={(item) => <Badge>{item}</Badge>}
+            />
+          ),
+        },
+        __typename: edge?.node?.__typename,
+      },
+    }));
 
   if (error) {
     if ((error as any).networkError?.statusCode === 403) {
@@ -119,17 +125,17 @@ function GlobalPermissions() {
   }
 
   if (isLoading) {
-    return (
-      <LoadingIndicator
-        message="Retrieving global permissions..."
-        className="h-[calc(100vh-13rem)]"
-      />
-    );
+    return <LoadingIndicator message="Retrieving groups..." className="h-[calc(100vh-13rem)]" />;
   }
 
   if (!permission?.view.isAllowed) {
     return <UnauthorizedScreen message={permission?.view?.message} />;
   }
+
+  const globalRefetch = () => {
+    queryClient.invalidateQueries({ queryKey: roleManagerQueryKeys.all });
+    refetch();
+  };
 
   return (
     <>
@@ -139,7 +145,7 @@ function GlobalPermissions() {
             loading={isFetching}
             value={search}
             onChange={(e) => setSearch(e.target.value)}
-            placeholder="Search global permissions"
+            placeholder="Search groups"
             className="border-none focus-visible:ring-0"
             containerClassName="grow"
           />
@@ -156,20 +162,20 @@ function GlobalPermissions() {
         <Table
           columns={columns}
           rows={rows ?? []}
+          className="border-0"
           onDelete={(data) => setRowToDelete(data.values)}
           onUpdate={(row) => {
             setRowToUpdate(row.values);
             setShowDrawer(true);
           }}
-          className="border-0"
           permission={permission}
         />
 
-        <Pagination count={data && data[GLOBAL_PERMISSION_OBJECT]?.count} />
+        <Pagination count={data && data[ACCOUNT_GROUP_OBJECT]?.count} />
       </div>
 
       <ModalDeleteObject
-        label={schemaKindName[GLOBAL_PERMISSION_OBJECT]}
+        label={schemaKindName[ACCOUNT_GROUP_OBJECT]}
         rowToDelete={rowToDelete}
         isOpen={!!rowToDelete}
         onOpenChange={(open) => !open && setRowToDelete(null)}
@@ -181,7 +187,7 @@ function GlobalPermissions() {
           title={
             <SlideOverTitle
               schema={schema}
-              currentObjectLabel={rowToUpdate?.identifier?.value ?? "New"}
+              currentObjectLabel={rowToUpdate?.name?.value ?? "New"}
               title={`${rowToUpdate ? "Update" : "Create"} ${schema.label}`}
               subtitle={schema.description}
             />
@@ -191,7 +197,7 @@ function GlobalPermissions() {
           onClose={() => setRowToUpdate(null)}
         >
           <ObjectForm
-            kind={GLOBAL_PERMISSION_OBJECT}
+            kind={ACCOUNT_GROUP_OBJECT}
             currentObject={rowToUpdate}
             onCancel={() => {
               setRowToUpdate(null);
@@ -210,5 +216,5 @@ function GlobalPermissions() {
 }
 
 export function Component() {
-  return <GlobalPermissions />;
+  return <Groups />;
 }
